fix build by renaming the local Rope module in Polyrope
Recent versions of the Core library now export a module named Rope,
which conflicts with the Rope module implemented as part of Myth's
library. This means that compiling Myth fails by default on users with
an up-to-date OCaml environment.

Rather than adding an upper bound on the supported Core version,
I thought it would be nice to allow Myth development to continue even
with recent Core version, so I renamed the local Rope module into
Polyrope -- the name reflects that it is not, like Core.Std.Rope is,
restricted to character ropes.

(I think it would still be a good idea to specify version bounds for
known-good Core versions in the opam file and maybe in the README, so
that non-expert users can fix the issue by installing the recommended
versions. On my machine I could easily test that 113.33.02 and
113.33.03 both work, using OCaml 4.04, 4.03 and 4.02.)